temporary fix for IPsec UE server not responding
The server socket for IPsec is opened for listening and accepting
connections. The socket descriptor for relying should be the one
returned by accept, The listening descriptor returns 'broken pipe'
when attempt to write on it is made

TODO: The processing of the accept in the SIP layer seems to be
performed after the reception of the data. The order should be fixed.